Move Morton key code from traversal.h to morton_key.h
