+++ include/morton_key.h
@@ -188,6 +188,26 @@ namespace exafmm {
       }
       return level;
     }
+
+    //! Get 3-D index from key
+    ivec3 getIndex(uint64_t key) {
+      int level = -1;                                           // Initialize level
+      while( int(key) >= 0 ) {                                  // While key has level offsets to subtract
+	level++;                                                //  Increment level
+	key -= 1 << 3*level;                                    //  Subtract level offset
+      }                                                         // End while loop for level offsets
+      key += 1 << 3*level;                                      // Compensate for over-subtraction
+      level = 0;                                                // Initialize level
+      ivec3 iX = 0;                                             // Initialize 3-D index
+      int d = 0;                                                // Initialize dimension
+      while( key > 0 ) {                                        // While key has bits to shift
+	iX[d] += (key % 2) * (1 << level);                      //  Deinterleave key bits to 3-D bits
+	key >>= 1;                                              //  Shift bits in key
+	d = (d+1) % 3;                                          //  Increment dimension
+	if( d == 0 ) level++;                                   //  Increment level
+      }                                                         // End while loop for key bits to shift
+      return iX;                                                // Return 3-D index
+    }
   }
 }
 #endif
